Service추가

+++ src/main/java/com/example/shop/common/ResponseEntity.java
@@ -3,9 +3,9 @@ package com.example.shop.common;
 public class ResponseEntity<T> {
     private final int status;
     private final T data;
-    private final int count;
+    private final long count;
 
-    public ResponseEntity(int status, T all, int count) {
+    public ResponseEntity(int status, T all, long count) {
         this.status = status;
         this.data = all;
         this.count = count;
